requirements.txt
python~=3.12


numpy~=1.26.4
pandas~=2.2.3
scipy~=1.14.1
matplotlib~=3.9.2
seaborn~=0.13.2
scikit-learn~=1.5.2
statsmodels~=0.14.4
joblib~=1.4.2

tensorflow-cpu==2.17.0
tensorflow-probability~=0.24.0
miceforest~=6.0.3
MissForest~=4.2.3

yfinance~=0.2.46
beautifulsoup4~=4.12.3
lxml~=5.3.0
openpyxl~=3.1.5
tqdm~=4.67.1

# TensorFlow ecosystem utilities, should already by installed with tensorflow-cpu
tensorboard~=2.17.1
absl-py~=2.1.0
protobuf~=4.25.5